Adapted AJ's Volume object to work with ApiActions
