+++ src/test/java/org/launchcode/demo/LittleOnlineLibraryApplicationTests.java
@@ -1,7 +1,10 @@
 package org.launchcode.demo;
 
 import org.junit.jupiter.api.Test;
+import org.launchcode.demo.models.ApiActions;
 import org.springframework.boot.test.context.SpringBootTest;
+
+import java.io.IOException;
 
 @SpringBootTest
 class LittleOnlineLibraryApplicationTests {
@@ -10,4 +13,9 @@ class LittleOnlineLibraryApplicationTests {
     void contextLoads() {
     }
 
+    @Test
+    void apiQuery() throws IOException {
+        String SearchResult = ApiActions.ApiSearch("Dune");
+        System.out.println(ApiActions.ParseResults(SearchResult));
+    }
 }
